refactor(super-editor): share block sdt navigation helpers
